Added New functions
checkSelectedTrue => A function that will check if the currently selected option matches the current option that is running. Will be used to exit out of animation loops.

getCurrentRadioOption => A function that gets all current radio options and returns whatever one is currently active. The code already exists I simply made it into a function.